Revise TBlockSort
Converted TBlockSort from a static into a normal class. This change was
made to avoid having to pass data to be sorted, and its size, to every
method. This information is now passed to a constructor, validated using
assertions and assigned to fields. Other methods were changed to access
the fields instead of having the information passed as parameters.

Made all parameters to methods const.

Fixed error in how memory allocation failures are trapped and reported.

Simplified calculation of comparison length in TBlockSort.Compare, using
new "inline if" feature of Delphi 13.

Modified the calling code in TDiffer.MakeDiff to instantiate & free a
TBlockSort object instead of calling its static methods.